Manually parse. Allow parsing up to a stop character.
Provides a nominal performance improvement.

Adds a `parseUntil` method that parses up to a user provided stop character (presumably either the single or double quote in most cases). The stop character can be escaped with a backslash as expected.

`parseUntil` returns an object with `end` and `value` properties. `end` is the position of the found stop character (or string.length if never found).